+++ src/include/Protocol/ServerCapabilities.hpp
@@ -116,13 +116,29 @@ struct ServerCapabilities
     std::optional<DocumentLinkOptions> documentLinkProvider = std::nullopt;
     bool colorProvider = false;
     bool renameProvider = false;
+    /**
+     * The server provides folding provider support.
+     *
+     * @since 3.10.0
+     */
+    bool foldingRangeProvider = false;
     bool inlayHintProvider = false;
     std::optional<DiagnosticOptions> diagnosticProvider = std::nullopt;
+    /**
+     * The server provides workspace symbol support.
+     */
+    bool workspaceSymbolProvider = false;
+    /**
+	 * The server provides call hierarchy support.
+	 *
+	 * @since 3.16.0
+	 */
+    bool callHierarchyProvider = false;
     std::optional<SemanticTokensOptions> semanticTokensProvider = std::nullopt;
     std::optional<WorkspaceCapabilities> workspace = std::nullopt;
 };
 NLOHMANN_DEFINE_OPTIONAL(ServerCapabilities, positionEncoding, textDocumentSync, completionProvider, hoverProvider, signatureHelpProvider,
     declarationProvider, definitionProvider, typeDefinitionProvider, implementationProvider, referencesProvider, documentSymbolProvider,
-    codeActionProvider, documentLinkProvider, colorProvider, renameProvider, inlayHintProvider, diagnosticProvider, semanticTokensProvider,
-    workspace);
+    codeActionProvider, documentLinkProvider, colorProvider, renameProvider, foldingRangeProvider, inlayHintProvider, diagnosticProvider, workspaceSymbolProvider,
+    callHierarchyProvider, semanticTokensProvider, workspace);
 } // namespace lsp